requirements.txt
requests
cryptography
pylint
rope
python-dotenv